requirements.txt
flask>=3.0.0
flask-restx>=1.3.0
kafka-python>=2.0.2
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.0
colorama>=0.4.0
tenacity>=8.0.0
pybreaker>=1.0.0
pytest>=8.0.0
pytest-flask>=1.3.0
